review changes: don't use big Util class & avoid asserts
